requirements.txt
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.1
cssselect2==0.3.0
defusedxml==0.6.0
drawSvg==1.6.1
imageio==2.9.0
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
plotly==4.9.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
retrying==1.3.3
six==1.15.0
tinycss2==1.0.2
webencodings==0.5.1
